backend/requirements.txt
aiohttp==3.7.3
aiohttp-devtools==0.13.1
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
devtools==0.6.1
gunicorn==20.0.4
idna==2.10
idna-ssl==1.1.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
multidict==5.1.0
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pkg-resources==0.0.0
plotly==4.14.1
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.54.1
typing-extensions==3.7.4.3
watchgod==0.6
xgboost==1.3.1
yarl==1.6.3
